Rework fixed size key generator to allow multiple widths
Before this change, the fixed length key generator made a list of keys
that used a constant number of unique values per digit. This constant
was the same for all digits in the key.

For example, with a constant of 2, it would generate 3 unique values
for each digit: 0, 128, and 255.

Changing this to an array allows tests to design key distributions
which has density at different places. For example, we could specify
an array of [128, 8, 2] to get a key of length 3 with a dense first
digit but that is very sparse in the lower region of the array.