Add memory test for key distribution with large prefixes
This new memory tests expands a fixed length key into a larger size
with guaranteed large runs of prefixes, testing the prefix compression
logic.

This test was supposed to be roughly comparable with the fixed length
memory test. The fixed length one used 306 keys, while the large prefix
test used 294 keys. The fixed length keys were of length 3, the large
prefix keys were of length 6.

The large prefix test outcome had a smaller `max_blocks` value, possibly
because the keys were concentrated into fewer inner nodes due to prefix
compression. The `max_bytes` value of the large prefix test was 4.5
times larger than the fixed length key. I think the cause of this is due
(1) because the key was 2x the size, and (2) because the key is
duplicated between the leaf nodes and the inner nodes + prefixes. These
explanations are mostly hypothesis.