WIP: Rework prefix storage to store limited bytes
This change removes the use of smallvec and instead uses a fixed size
buffer. This strategy represents up to 8 bytes of the prefix in memory
and then the rest of the prefix is implicit. This is how the original
ART paper describes the prefix implementation.

Currently this implementation is incomplete, there are some specific
cases in the insert function which are broken.