Rename allocate/deallocate and write key prefix functions
**Description**
 - Rename `NodePtr::{de}allocate_node` to `{de}allocate_node_ptr`
 - Rename `Header::write_prefix` to `extend_prefix`

**Motivation**
I thought these names were more accurate. For the key prefix method,
`extend_prefix` is more specific and conveys that it adds to the end.

For the allocate/deallocate methods, I thought it was important to
convey that it was allocate a node onto the heap, not creating the
node originally.

**Testing Done**
 - `cargo test`
 - `cargo build --all`
 - `cargo fuzz build`