Implement range constructors for inner node iterators
**Description**
 - Add new constructors named `range` to all the `InnerNode` iterator
   types. These constructors take an additional input of the `RangeBounds`
   type, and constrain the set of children output to those whose key
   bytes fall in the given range.
 - Add some README updates for getting fuzzing code coverage
 - Move "should shrink" method into `NodeType` as inherent method.
 - Move iteration tests into own file, add additional unit tests
   which cover many different cases of range types.

**Motivation**
This change is a prerequisite to creating some tree iterators that take
a range argument. My basic idea is that eventually we'd have some sort
of tree range iterator constructor that will move down the tree and
constrain the stack of iterators based on the relevant key bytes at that
level.

**Testing Done**
 - `cargo build --tests --examples --benches`
 - `cargo test`
 - `cargo miri test`
 - `cargo clippy`