Implement most of tree map wrapper over raw nodes API
**Description**
Implements most of issue #12.

This change introduces a new module that contains a safe wapper over
the unsafe node APIs. It is based directly on the API of the BTreeMap
that is present in the standard library.

Some function are not implemented, currently they are marked with
`todo!()`.

Also added in a fuzz target which covers the integration of the raw
node APIs into the safe wrapper.

**Motivation**
This change is needed so that this crate can be useful to people
without forcing them to use lots of unsafe functions. This wrapper
takes care of enforcing safety conditions and matches most expecations
around which APIs a map container should provide.

**Testing Done**
 - `cargo test`
 - `cargo miri test`
 - `cargo clippy`
 - `cargo fuzz run fuzz_tree_map_api`
 - `cargo doc`