Replace unstable std lib functions with local versions
**Description**
Work towards issue #20

For each unstable standard library function we depend on, copy the
implementation and documentation into the local crate and replace
all usages of it.

Added links to tracking issues so we can eventually swap back.

**Motivation**
This change is needed so that this crate can compile using the
stable rust distribution.

The downside to this change is that we will no longer receive fixes
or other patches that the stdlib does on their version. If they
uncover an issue, we won't know about it. Going to mitigate this
by subscribing to each tracking issue.

**Testing Done**
 - `cargo test`
 - `cargo miri test`
 - `cargo clippy`