Make crate compile and pass tests on aarch64 platform
**Description**
Use conditional compilation throughout the crate to make it compile and
run on aarch64 platform.

Mainly removed use of perfcnt crate and associated performance counters
on aarch64, revert to wall time for benchmark measurement.

One of the tagged pointer test started failing because u128 has a
different alignment on aarch64 versus x86. There is a rust-lang issue:
https://github.com/rust-lang/rust/issues/54341

**Motivation**
This change is needed so I can develop this crate on an M1 mac.

**Testing Done**
 - `cargo test`
 - `cargo build --all-targets`
 - `cargo miri test`