Remove use of slice_ptr_get unstable feature
**Description**
Replace use of `NonNull::<[T]>::get_unchecked_mut` function call with
a copy from the standard library.

This function is used to index into a pointer to a slice (`*const [T]`)
and return a pointer to an element (`*const T`) without doing bounds
checking.

**Motivation**
This change is a partial solution towards issue #20. The only remaining
unstable feature used is `hasher_prefixfree_extras`.

**Testing Done**
 - `cargo test`
 - `cargo miri test`
 - `cargo fmt`
 - `cargo clippy`