Add cargo feature to enable unstable nightly functions
**Description**
Add a new cargo feature called `nightly` that can enable or disable the
use of unstable nightly features like `slice_ptr_get`.

**Motivation**
This change is useful so that we can continue to test with full features
on a nightly CI environment, while developing the crate mainly for
stable.

This also allows us to use a better implementation for some of the
nightly features like `hasher_prefixfree_extras`. If stdlib updates
their implementation we would otherwise not benefit.

**Testing Done**
 - `cargo test && cargo test --features nightly`
 - `cargo clippy && cargo clippy --features nightly`
 - `cargo fmt`