Implement a SIMD version of the child-pointer lookup
**Description**
Implement a SIMD optimized version of the child pointer lookup operation
for `InnerNodeCompressed` types. When the number of children is greater
than or equal to a threshold (currently 16), the the lookup function
uses SIMD instruction to look for the haystack byte in parallel.

Also alphabetically sorted the nightly features and removed
`maybe_uninit_write_slice` which was unused.

**Motivation**
This could possibly make the overall tree search function faster by
increasing the speed of the InnerNode16 lookup.

**Testing Done**
 - `cargo test && cargo test --features nightly`