Implement split operation for inner nodes
**Description**
Add a new method to the InnerNode trait which allows splitting the node
at a specific key byte, keeping all child pointers less than the key byte,
moving all child pointers greater than the key byte to a new node.

**Motivation**
This is in preparation for a new `split_off` operation that will take an
entire tree and split it in half at a specific key.

**Testing Done**
 - Added new unit tests for all inner node types testing their split
   implementations.
 - `cargo test && cargo miri test && cargo clippy && cargo fmt`