Split stable versus nightly for memory usage tests
**Description**
Add a check for the `nightly` feature flag to change the expected sizes
in different tests.

This seems like a short-term solution because the nightly changes will
eventually get promoted to stable and then this change will need to be
removed.

I found a possibly related [rust-lang issue #104807][rust-issue-104807]
, maybe the fix for this will change the nightly enum sizes back to
their previous values

**Motivation**
CI was breaking on the nightly and miri tests

**Testing Done**
 - `cargo +nightly test --features nightly` passes
 - `cargo +nightly miri test` passes
 - `cargo +nightly test` fails

[rust-issue-104807]: https://github.com/rust-lang/rust/issues/104807