Add a tree visitor which verifies expected structure
**Description**
Add a new visitor which traverses a tree and verifies a couple of
conditions to hold. Namely:
 - No loops in the tree
 - Each inner node has the right number of children
 - The built-up key prefix from the inner node search path matches
   the prefix of the actual key in the leaf.

If any of these conditions are violated a single error is returned with
some descriptive fields.

**Motivation**
Most of the tree operations have an implicit (in some cases explicit)
safety condition that requires the tree to be "well-formed". Adding
this visitor enables us to check the condition easily in unit tests or
fuzz tests.

**Testing Done**
 - `cargo test`
 - `cargo clippy`