Replace prev alignment fix with required alignment minimum
**Description**
Replace previous fix for the K/V alignment issue with a different fix
that instead encodes the required minimum alignment/number of bits as a
const generic on the type. The minimum number of bits if checked against
the number of provided bits at compile time and will cause a failure if
it does not have enough.

**Motivation**
This is a better fix than the previous because it does not require us to
sacrifice more allocations/memory usage to work with large alignments.

**Testing**
 - Previous test which ensures functionality with large alignments still
   still passes
 - All other tests pass (Miri was weird about deallocating the tagged
   pointers)