feat: Add safe conversions of integers to bytestrings
**Description**
Add a trait and some convenience types so that signed and unsigned
integers are converted to bytestring using a method that will preserve
their natural ordering.

**Motivation**
These conversions are useful because that allows us to use these integer
types as keys in the `TreeMap` and be able to access them with range
queries.

**Testing Done**
 - Added unit tests on the conversions making sure that for some numbers
   the safety variants hold