Separate header definition into own module
**Description**
Isolate the `Header` type into a separate module with separate tests.

I also  modified the word count example to print out the tree stats,
just for fun.

**Motivation**
This change is useful because I'd like to change the internal
representation of the `Header` type and making the internals private
makes that easier in the future

**Testing Done**
 - `cargo test`
 - Ran profiler: https://share.firefox.dev/4c5jo77