Remove `tinyvec` optimization
**Description**
The ART paper describes the header structure as only retaining up to
8 bytes of the inner node prefix before it starts to represent it
implicitly as a length instead.

I couldn't figure out how to implement the implicit code path on
insert, so I decided to try and go halfway and keep up to 8 bytes
inline, and then spill to the heap once it grew past that.

However, I recently looked at a profile of the code and the `tinyvec`
code, specifically the part which switches between inline vs heap
implementations, was very hot. So I decided to try removing it, and
I was able to get a 5% reduction in instructions (on average).

**Testing Done**
`cargo test`