Implement basic range iterator
**Description**
 - Implement the basic range iterator which will returns tuples of
   `(key, value)`.
 - Add the `TreeMap::range` function and a doctest example
 - Remove some dead code comments that were lying around

**Motivation**
This is the central feature that I wanted to implement, the ability
to query a sub-section of the `TreeMap` using the native key types.

This commit only implements the `Range` iterator, not the `RangeMut`
iterator so that the initial commit is devoid of `macro_rules` stuff.

This commit also includes some useful machinery that I think could
be applied to the `Prefix*` iterators, namely the
`find_terminating_node` function.

**Testing Done**
Full nightly test. I was not able to run a fuzzer because of some
difficulties setting it up.