Re-implement specialized clone operation
**Description**
The specialized clone based on `deep_clone` was removed because it
did not account for the leaf node linked list. This commit
re-implements `clone()` using a non-recursive algorithm that should
be more efficient than creating an empty tree and inserting elements.

This commit is needed, otherwise it would be a significant perf hit
to add the leaf node linked list.

This commit also removes the `deep_clone` functions and associated
stuff.

**Testing Done**
`./scripts/full-test.sh nightly`