Ensure all iterators and map-related types are Send/Sync
**Description**
Add tests and unsafe impls to ensure that the `TreeMap` and related
types are `Send` and `Sync`.

**Motivation**
This is needed to match the interface of the `BTreeMap` and to make
sure we don't regress on this behavior between releases.

The `unsafe impl`s are not my favorite way to implement this, I would
have preferred that the building block types were naturally `Send` or
`Sync`, that wasn't easily done. Maybe a future improvement to remove
those.

**Testing Done**
`./scripts/full-test.sh nightly`