Add optional support for custom TreeMap allocator (#45)
* Add optional support for custom TreeMap allocator

**Description**
 - Modify the TreeMap interface so that it exposes a new generic
   parameter for the allocator, defaulting to the Rust global
   allocator.
 - Modify the internal tree operations to use the allocator stored
   in the tree for insert/delete/clone/dealloc operations
 - Add a new feature flag and optional dependency to consume the
   `Allocator` trait on the stable toolchain, also integrate with the
   `nightly` feature flag + toolchain to use that trait.
 - Also fix a bunch of clippy lints that are coming in on the latest
   nightly toolchains
   - Converting explicit lifetimes to `'_`

Note: The current OOM behavior is to panic. I may change this in the
future to support an explicit `Err(...)`, but not to start at least.

**Motivation**
I think custom allocators are a very useful feature for making the
the library more flexible and useful in `no_std` environments, or
other non-standard scenarios. I was even able to experiment with
using `bumpalo` as the custom allocator, which I think is neat.

**Testing Done**
I added some unit tests which create various custom allocators and
use those to create/update/query/delete from `TreeMap`.

* Use nightly function replacement to fix stable build

Also fix some clippy lints.