Extend fuzzer to catch well-formed checker bug for N256
**Description**
 - Extend the fuzzer so that it can easily insert and delete many entries
 - Fix a bug in the well-formed checker that would incorrectly report an
   error if a Node256 actually had 256 children.

**Motivation**
I noticed in a previous fuzzing run coverage report that the code
branches for Node256 were largely uncovered. I figured this was because
its difficult for the fuzzer to insert 256 non-overlapping entries.

Adding explicitly actions for insert and deleting many nodes makes it
easier for the fuzzer to explore these code paths.

**Testing Done**
Added regression unit test which now passes.