Fix a well-formed checker bug for N256
**Description**
Fix a bug in the well-formed checker that would incorrectly report an
error if a Node256 actually had 256 children.

**Motivation**
I noticed in a previous fuzzing run coverage report that the code
branches for Node256 were largely uncovered. I figured this was because
its difficult for the fuzzer to insert 256 non-overlapping entries.

This is a cherry-pick of 6d37b4d82cd6c0197d5b495d44dd0f38feef4025.

**Testing Done**
Added regression unit test which now passes.