Fix tree corruption bug in entry module
**Description**
The `entry` module contains an entry type for vacant and another type
for occupied entries. Previously there was a bug that could happen if
the user used the `entry`/`try_entry` API, called `insert_entry` (to go
from vacant to occupied), then called `remove` (to go from occupied to
vacant).

The problem was that the parent pointers that were used to insert the
new leaf node were invalidated after the insert, but were still being
used for the delete call (within `remove`).

This commit fixes the issues by:
 - Introducing some unsafe around the `apply_insert*` and
   `apply_delete*` functions, so that callers understand the effects.
 - Return information from the `InsertPoint::apply` and
   `DeletePoint::apply` that is sufficient to update parent pointers in
   case they are kept across an insert or delete.
 - Add a function to the entry module that will fixup the parent
   pointers using the newly returned information from the insert call.
 - Add a lot more tests

**Motivation**
I want the entry module and related functions to work correctly and not
corrupt the tree state.

Fixes issue #48.

**Testing Done**
8+ hours of fuzzing and added unit tests.