Simplify `ExtractIf` implementation
**Description**
 - Redo the `ExtractIf` implementation so that it uses the same leaf
   linked-list to traverse the tree (as other iterators do). Instead of
   maintaining parent and DFS state, the iterator just deletes leafs
   using the normal `TreeMap::remove` API.
 - Remove some now unused code that was associated with the
   previous version of `ExtractIf`
 - Extend the `ExtractIf` implementation to implement
   `DoubleEndedIterator` and support range arguments

**Motivation**
I wanted to support the range and double-ended features, but making
those work with the stack-based implementation was going to be very
tricky.

Also, there was a downside to the stack method, in that it used memory
proportional to the heigh of the tree during iteration. In the very
worst case (a highly skewed tree), it would end up holding pointers to
every single node in the tree. This new implementation uses constant
amount of memory since it is traversing a linked-list.

The new implementation is slower for removing, but faster for
traversing. I'll include benchmark comparison in the PR comments. I
think this is an acceptable tradeoff in this case for the added
features.

**Testing Done**
Added more unit tests, `./scripts/full-test.sh nightly`