Rename `test_common` module and place behind feature flag
**Description**
Before this testing module was a hidden part of the public API, this
change makes it so that users would need to opt in.

I learned from https://github.com/rust-lang/cargo/issues/2911 that
you can take a `dev-dependency` recursively with extra features
enabled, this allows the feature to be off by default and still used
in the dev dependencies.

**Testing Done**
Unit tests already exercise this code.