Make InnerNode an unsafe trait
All types that implement `InnerNode` must  be `repr(C)` and have a
`Header` as their first field. This allows other unsafe code to rely on
casting a pointer to any inner node to a pointer to a `Header`. This
commit formalizes this requirement by marking `InnerNode` as an
`unsafe trait`.

Also change `leaf_node_ptr.read()` to `leaf_node_ptr.as_ref()` to avoid
an unnecessary copy.