Split `InnerNode` trait
**Description**
Split `InnerNode` trait into `InnerNode` and `InnerNodeCommon`.
`InnerNodeCommon` has all the functions related to accessing and
modifying the inner node. `InnerNode` is a sub-trait that adds on info
about the growth order of the inner nodes and defines a specific
`NodeType`.

**Motivation**
I made the split so that I could implement `InnerNodeCommon` for
`InnerNodeSorted<..., SIZE>` and `InnerNodeIndirect<..., SIZE>` and
have it be generic over `SIZE`. That wouldn't be possible if I had done
that before the split, because the `NodeType` for each `InnerNode` impl
needs to be specific to the `SIZE`.

**Testing Done**
`./scripts/full-test.sh nightly`